# Add faster method to calculate pscore item position