@usaito 2021-07-02
@fullflu Thanks!

- What do you mean by `policy_value`? We often use policy value to express the performance of a policy, but you use this word in a different meaning..., so I think you'd better rephrase this word and use `policy_softmax_value` or `softmax_value`.
https://github.com/fullflu/zr-obp/blob/94e9f77c3d117e1eaf9fc9bd11500da64c2fe572/obp/dataset/synthetic_slate.py#L362
https://github.com/fullflu/zr-obp/blob/94e9f77c3d117e1eaf9fc9bd11500da64c2fe572/obp/dataset/synthetic_slate.py#L372

- clip_logit_value?
https://github.com/fullflu/zr-obp/blob/94e9f77c3d117e1eaf9fc9bd11500da64c2fe572/obp/dataset/synthetic_slate.py#L734