@fullflu 2021-07-11
> What do you mean by policy_value? We often use policy value to express the performance of a policy, but you use this word in a different meaning..., so I think you'd better rephrase this word and use policy_softmax_value or softmax_value.

Thank you, I will use `policy_softmax_value`.

> clip_logit_value?

That's right, thank you!